Nyxian Voice, Level 3 nyxian cameo, WallRunnable Wall Mat update, Decals galore
Ok so I did a bunch of shit in level 3, specifically made it so when you leave the room a nyxian spawns and walks past and the disappears in a puff of smoke after passing, leaving behind a bunch of crazy effects. It also whispers.

I also made a material for the wall runnable walls to have so we can tell they're wallrunnable when we get close to them. the material is a fucking headache because I can't figure out how to scale down the effect I want.