Dialogue trigger obj, Fixed Searinglight1, Made Rust_2 material for a floor, Make glass floor breakable
I made an Dialogue Trigger thing that pops up text when you pass through it, similar to the text of the buttons.

I realized I'd fucked up the specifics of the searinglight so I fixed that.

I made a material for the crucible floor based on the rust materials

And I made the glass floor at the center breakable of the start of the skylight section breakable by stomping.