Vampire BP Issue, Solar Flare/Flashbang input change
Solar Flare now stores the charge when you let go, which can greatly reduce your flashlight power. but once you turn it on again but hitting F an additional time, the Solar Flare Fires, with the bonus of leaving the flashlight on but having the light intensity return to normal slowly over time.

You can also use the flash on the vampires, which will damage them, but not kill them, as that takes several flashes. If you do flash it enough to kill it, it teleports away to a random spot in range, which can include right behind you.

also, for some reason, I don't know if this is just me, but every time I try to compile the vampire BP my entire UE5 freezes and I have to end task. so if you mess with the Vampire BP, save before you try to compile, as thats the only way to actually make changes to that BP