Grapple Lock Physic Update, Added doors to a few levels
Ok so I updated the physics of the Lock Grapple so now its always using the normal bungee physics, but its tension is lerped dependant on your speed, so if your going too slow the lock slowly stops. if you hold shift, the lock is always on.

I also updated some geometry