Junkfield Dialogue, entry room remodel, distance based Pickups
Ok so The Junkfield now has dialogue informing you what to do

I remodelled the room a bunch the way its designed is the elevator we enter from is part of the tram system/freight elevator, and we enter on foot onto a waiting area. we're basically in a tram station with the area to the right of the entry being implied to be where the tracks area.

also I made special pickup light fragment items that only appear if your within 1000 units.