Maze flashlight aquisition now has dialogue, flashlight prop physics fixes
Ok so I juust set it up so the flashlight can't bounce away anymore, and I set up dialogue for before and after the flashlight appears.

Also the trees now look really cool and have an earthquake effect when they appear. there's SUPPOSE to be a particle effect but it won't show up due to the angle anyways it 5 am so I'm going to bed GN