Maze Map Now has Shortcuts
Ok so the Maze now has its shortcuts in it, and I set it up so each tree that gets flashed sets a checkpoint there for you to respawn at. while there is a nonzero chance you can get spawncamped, the flashcharge doesn't reset, so you can tank through it and flash the vampire away.