Adding More trees and the Settings menu.
I've added two new trees to the project and the Settings Menu which is a "work in progress".

Planning to work on changing the background music, sound effects, and probably the voice (I'm not sure yet).

Planning to add a custom slider with material that I found on YouTube as an idea and I'm going to use that.

More to come.