Creating Custom Slider with Materials for the slider and testing the audio.
I am done creating the custom slider with Materials for the slider from the tutorial that I found on YouTube.

Next is working on testing the Background music volume and sound effects by using the settings menu for testing purposes (Not done on it yet).

Still a work in progress as I have to figure out how to put the background music in the main menu and during the level and figure out how to go to the settings with buttons, and go back to the menu and/or level from the settings.