Watcher Eye Update, 360 Lock fix, Particle Tweak
Ok so the watcher eye now looks towards the player if they're in its LOS. if they're hiding/not in line of sight, it'll wait for a bit on the last place the player was visible, then look around a bit, then give up and revert to neutral.

I also finally fixed the lock grapple so it doesn't activate if your lock grappling while already holding shift, you have to fully release and press the button again.

Finally I adjusted how the damage is visualized, now there's an overlay that appears and gets more solid the more the damage, as well as finally figuring out how to cull particles that are RIGHT up to the camera, so its no longer blinding.