SINGLE INCREMENT DAMAGE CAPABILITY, Damage Turret, Bounce turret
Ok so there's 2 new turrets, one that bounces you if your hit by the shot/nearby when it moves past, and the other one that deals damage if your hit by it.

To this end, I added the "Deal Damage" event to the BP thirdpersoncharacter to allow us to add chunks of damage in sections, with a retriggerable second delay, so when you stop taking damage for x seconds, it restores. I had to use a different restore for it because FOR SOME REASON THE FIRST TIME THE RESTORE DOESN'T TRIGGER.