Various Fixes and Tweaks, Fixed Demo Checkpoints, Fixed Repeating Dialogue in LV 1
I restored the ORIGINAL checkpoint system that changes the place you respawn to the last place you hit the floor, because every time a respawn is triggered it resets that value to the checkpoints position. So the only way to reset to last ground touch is using the "Demo Killbox".

The repeating dialogue was caused by, for some reason, the question not registering the anwsers correctly. I just switched it out with a fresh version and it worked.

Made the one hall in level 1 better.
Fixed an issue with the respawn groundboxes for level 7